BERTResearch/requirements.txt
praw
pandas
python-dotenv
tqdm